feat(initial): initial version of containers
This is probably as far as I'm going to be able to get. After much work to try
and get everything working, it appears that there is a library that isn't
available for ARM architectures, `c42archive.so` that is going to keep this
from going any further. It was a nice experiment, but in the end, this project
isn't going to be useful.

It probably had a limited lifespan anyway as Code 42 is moving to a new
Electron based frontend in the near future. That would bring other chagnes that
would probably break the application on ARM architectures.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>